ref: Split platform module to avoid circular deps
