+++ packages/core/platform/device/index.ios.ts
@@ -1,8 +1,4 @@
-﻿/* tslint:disable:class-name */
-
-import { platformNames } from './common';
-import { ios } from '../utils';
-export * from './common';
+import { platformNames } from '../common';
 
 type DeviceType = 'Phone' | 'Tablet' | 'Vision';
 
@@ -87,48 +83,7 @@ class DeviceRef {
 	}
 }
 
-class MainScreen {
-	private _screen: UIScreen;
-
-	private get screen(): UIScreen {
-		if (!this._screen) {
-			// NOTE: may not want to cache this value with SwiftUI app lifecycle based apps (using NativeScriptViewFactory) given the potential of multiple scenes
-			const window = ios.getWindow();
-			this._screen = window ? window.screen : UIScreen.mainScreen;
-		}
-
-		return this._screen;
-	}
-
-	get widthPixels(): number {
-		return this.widthDIPs * this.scale;
-	}
-	get heightPixels(): number {
-		return this.heightDIPs * this.scale;
-	}
-	get scale(): number {
-		return this.screen.scale;
-	}
-	get widthDIPs(): number {
-		return this.screen.bounds.size.width;
-	}
-	get heightDIPs(): number {
-		return this.screen.bounds.size.height;
-	}
-
-	public _updateMetrics(): void {
-		// UIScreen handles the update on iOS
-	}
-}
-
 export const Device = new DeviceRef();
 
 // This retains compatibility with NS6
 export const device = Device;
-
-export class Screen {
-	static mainScreen = new MainScreen();
-}
-
-// This retains compatibility with NS6
-export const screen = Screen;
